+++ control-core/src/main/kotlin/at/florianschuster/control/errors.kt
@@ -15,7 +15,7 @@ internal sealed class ControllerError(
         tag: String,
         action: String,
         cause: Throwable
-    ) : ControllerError("Mutator error in $tag, action = $action", cause)
+    ) : ControllerError(message = "Mutator error in $tag, action = $action", cause = cause)
 
     /**
      * Error during [Reducer].
@@ -28,5 +28,19 @@ internal sealed class ControllerError(
     ) : ControllerError(
         message = "Reducer error in $tag, previousState = $previousState, mutation = $mutation",
         cause = cause
+    )
+
+    /**
+     * Error during [EffectEmitter.emitEffect].
+     */
+    class Effect(
+        tag: String,
+        effect: String
+    ) : ControllerError(
+        message = "Effect error in $tag, effect = $effect",
+        cause = IllegalStateException(
+            "Capacity for effects has been reached. Either too many effects have been triggered " +
+                "or they might not be consumed."
+        )
     )
 }