+++ examples/android-github/src/main/kotlin/at/florianschuster/control/androidgithub/search/SearchView.kt
@@ -1,77 +1,81 @@
 package at.florianschuster.control.androidgithub.search
 
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.view.View
-import android.widget.EditText
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import at.florianschuster.control.androidgithub.R
+import at.florianschuster.control.androidgithub.databinding.ViewSearchBinding
+import at.florianschuster.control.androidgithub.showSnackBar
+import at.florianschuster.control.androidgithub.viewBinding
 import at.florianschuster.control.bind
 import at.florianschuster.control.distinctMap
-import at.florianschuster.control.androidgithub.R
-import at.florianschuster.control.androidgithub.databinding.ViewGithubBinding
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.sample
 import reactivecircus.flowbinding.android.widget.textChanges
 import reactivecircus.flowbinding.recyclerview.scrollEvents
 
-internal class GithubView : Fragment(R.layout.view_github) {
+internal class SearchView : Fragment(R.layout.view_search) {
 
-    private var binding: ViewGithubBinding? = null
-    private val requireBinding: ViewGithubBinding get() = requireNotNull(binding)
+    private val binding by viewBinding(ViewSearchBinding::bind)
+    private val viewModel by viewModels<SearchViewModel> { SearchViewModel.Factory }
 
-    private val viewModel: GithubViewModel by viewModels { GithubViewModelFactory }
-
-    private val repoAdapter = RepoAdapter { repo ->
-        startActivity(Intent(Intent.ACTION_VIEW, repo.webUri))
+    private val repoAdapter = SearchAdapter { repo ->
+        startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(repo.webUrl)))
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding = ViewGithubBinding.bind(view)
 
-        with(requireBinding.repoRecyclerView) {
+        with(binding.repoRecyclerView) {
             adapter = repoAdapter
             itemAnimator = null
+            addItemDecoration(DividerItemDecoration(requireContext(), LinearLayoutManager.VERTICAL))
         }
 
         // action
-        requireBinding.searchEditText.textChanges()
+        binding.searchEditText.textChanges()
             .debounce(SearchDebounceMilliseconds)
             .map { it.toString() }
-            .map { GithubViewModel.Action.UpdateQuery(it) }
+            .map { SearchViewModel.Action.UpdateQuery(it) }
             .bind(to = viewModel.controller::dispatch)
             .launchIn(scope = viewLifecycleOwner.lifecycleScope)
 
-        requireBinding.repoRecyclerView.scrollEvents()
+        binding.repoRecyclerView.scrollEvents()
             .sample(500)
             .filter { it.view.shouldLoadMore() }
-            .map { GithubViewModel.Action.LoadNextPage }
+            .map { SearchViewModel.Action.LoadNextPage }
             .bind(to = viewModel.controller::dispatch)
             .launchIn(scope = viewLifecycleOwner.lifecycleScope)
 
         // state
-        viewModel.controller.state.distinctMap(by = GithubViewModel.State::repos)
+        viewModel.controller.state.distinctMap(by = SearchViewModel.State::repos)
             .bind(to = repoAdapter::submitList)
             .launchIn(scope = viewLifecycleOwner.lifecycleScope)
 
-        viewModel.controller.state.distinctMap(by = GithubViewModel.State::loadingNextPage)
-            .bind(to = requireBinding.loadingProgressBar::isVisible::set)
+        viewModel.controller.state.distinctMap(by = SearchViewModel.State::loadingNextPage)
+            .bind(to = binding.loadingProgressBar::isVisible::set)
             .launchIn(scope = viewLifecycleOwner.lifecycleScope)
-    }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        binding = null
+        // effect
+        viewModel.controller.effects.onEach { effect ->
+            when (effect) {
+                is SearchViewModel.Effect.NetworkError -> {
+                    binding.root.showSnackBar(R.string.info_network_error)
+                }
+            }
+        }.launchIn(scope = viewLifecycleOwner.lifecycleScope)
     }
 
     private fun RecyclerView.shouldLoadMore(threshold: Int = 8): Boolean {
@@ -81,10 +85,5 @@ internal class GithubView : Fragment(R.layout.view_github) {
 
     companion object {
         const val SearchDebounceMilliseconds = 500L
-
-        internal var GithubViewModelFactory = object : ViewModelProvider.Factory {
-            @Suppress("UNCHECKED_CAST")
-            override fun <T : ViewModel?> create(modelClass: Class<T>): T = GithubViewModel() as T
-        }
     }
 }